Crate sphrs

A general purpose spherical/solid harmonics library in Rust.

Documentation: stable, main.

Types of spherical/solid harmonics

This crate supports these types of real and complex functions via the enums RealSH and ComplexSH:

Spherical

RegularSolid and IrregularSolid

Usage

Add this to your Cargo.toml:

[dependencies]
sphrs = "0.1.3"

Examples

Compute the complex spherical harmonic function of degree 2 and order 1 at (spherical) position (r = 1.0, theta = PI/4, phi = PI/4):

use sphrs::{ComplexSH, Coordinates, SHEval};
use std::f64::consts::PI;

let sh = ComplexSH::Spherical;
let degree = 2;
let order = 1;
let p = Coordinates::spherical(1.0, PI/4.0, PI/8.0);
println!("SH ({}, {}): {:?}", degree, order, sh.eval(degree, order, &p));

Compute all real SH up to 5th degree at (Cartesian) position (1, 0, 0):

use sphrs::{RealSH, HarmonicsSet, Coordinates};
let degree = 5;
let sh: HarmonicsSet<f64, _, _> = HarmonicsSet::new(degree, RealSH::Spherical);
let p = Coordinates::cartesian(1.0, 0.0, 0.0);
println!("SH up to degree {}: {:?}", degree, sh.eval(&p));

Acknowledgements

This crate is heavily inspired by Google’s spherical-harmonics library and follows the design documented here.

References

Robin Green, “Spherical Harmonic Lighting: The Gritty Details”

Spherical harmonics (Wikipedia)

Solid harmonics (Wikipedia)

License

Licensed under either of

Apache License, Version 2.0, (LICENSE-APACHE or http://www.apache.org/licenses/LICENSE-2.0)

MIT License (LICENSE-MIT or http://opensource.org/licenses/MIT)

at your option.

Contribution

Unless you explicitly state otherwise, any contribution intentionally submitted for inclusion in the work by you, as defined in the Apache-2.0 license, shall be dual licensed as above, without any additional terms or conditions.

Structs

Coordinates

Representation of coordinates.

HarmonicsSet

A set of spherical/solid harmonics up to a given degree

Enums

ComplexSH

Available types of complex spherical harmonics and solid harmonics

RealSH

Available types of real spherical harmonics and solid harmonics

Traits

SHCoordinates

SHCoordinates trait

SHEval

SH eval trait (TODO)

SphrsFloat

Trait alias to simplify common trait bounds

Functions

SH

Complex spherical harmonics

irregular_solid_SH

Complex irregular solid harmonics

real_SH

Real spherical harmonics (recursive implementation)

real_SH_hardcoded

Spherical harmonics. This will use the hardcoded functions if available and the recursive implementation otherwise.

real_irregular_solid_SH

Real irregular solid harmonics

real_regular_solid_SH

Real regular solid harmonics

regular_solid_SH

Complex regular solid harmonics

sh00

Hardcoded SH (l=0,m=0)

sh1n1

Hardcoded SH (l=1,m=-1)

sh1p1

Hardcoded SH (l=1,m=1)

sh2n1

Hardcoded SH (l=2,m=-1)

sh2n2

Hardcoded SH (l=2,m=-2)

sh2p1

Hardcoded SH (l=2,m=1)

sh2p2

Hardcoded SH (l=2,m=2)

sh3n1

Hardcoded SH (l=3,m=-1)

sh3n2

Hardcoded SH (l=3,m=-2)

sh3n3

Hardcoded SH (l=3,m=-3)

sh3p1

Hardcoded SH (l=3,m=1)

sh3p2

Hardcoded SH (l=3,m=2)

sh3p3

Hardcoded SH (l=3,m=3)

sh10

Hardcoded SH (l=1,m=0)

sh20

Hardcoded SH (l=2,m=0)

sh30

Hardcoded SH (l=3,m=0)