Crate sphrs

A (work in progress) general purpose spherical/solid harmonics library in Rust.

Documentation: stable, master.

Types of spherical/solid harmonics

This crate supports these types of real SH via the enum RealSHType:

Spherical

RegularSolid and IrregularSolid

TODO: complex SH

Usage

Add this to your Cargo.toml:

[dependencies]
sphrs = "*"

Example

Compute the sum of all real SH up to 5th order at position (1, 0, 0):

use sphrs::{RealSHType, RealHarmonics, Coordinates};
let order = 5;
let sh = RealHarmonics::new(order, RealSHType::Spherical);
let p = Coordinates::cartesian(1.0, 0.0, 0.0);
println!("SH up to order {}: {:?}", order, sh.eval(&p));

Acknowledgements

This crate is heavily inspired by Google's spherical-harmonics library and follows the design documented here.

References

Robin Green, "Spherical Harmonic Lighting: The Gritty Details"

Spherical harmonics (Wikipedia)

Solid harmonics (Wikipedia)

License

Licensed under either of

Apache License, Version 2.0, (LICENSE-APACHE or http://www.apache.org/licenses/LICENSE-2.0)

MIT License (LICENSE-MIT or http://opensource.org/licenses/MIT)

at your option.

Contribution

Unless you explicitly state otherwise, any contribution intentionally submitted for inclusion in the work by you, as defined in the Apache-2.0 license, shall be dual licensed as above, without any additional terms or conditions.

Re-exports

pub use crate::coords::*;

pub use crate::sh::*;

Modules

coords
Coordinates

sh
Spherical/solid harmonics

Structs

ComplexHarmonics
Complex spherical/solid harmonics

RealHarmonics
Real spherical/solid harmonics

Enums

ComplexSHType
Available types of complex spherical harmonics and solid harmonics

RealSHType
Available types of real spherical harmonics and solid harmonics

Traits

SphrsFloat
Trait alias to simplify common trait bounds